restructure python
