+++ django/joist/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'rest_framework.authtoken',
     'core',
+    'joist'
 ]
 
 MIDDLEWARE = [
